decoder/requirements.txt
Django~=3.0.8
mongoengine~=0.20.0
djangorestframework~=3.11.0
pymongo~=3.10.1